crawler/requirements.txt
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
cycler==0.10.0
dnspython==1.16.0
gevent==1.4.0
greenlet==0.4.15
idna==2.9
kiwisolver==1.1.0
matplotlib==3.1.1
mysql==0.0.2
mysql-connector-python==8.0.19
mysqlclient==1.4.6
numpy==1.17.2
parallel-ssh==1.9.1
paramiko==2.6.0
protobuf==3.6.1
pssh==2.3.1
pycparser==2.19
PyNaCl==1.3.0
pyparsing==2.4.2
python-dateutil==2.8.0
requests==2.23.0
scipy==1.3.1
six==1.12.0
ssh2-python==0.18.0.post1
urllib3==1.25.8
websockets==8.1
